Switched headless browser from PyQt5 to selenium
I could not get PyQt5 to play well with Binder, though it worked locally
in the IPython interpreter and in Jupyter notebooks (both in my normal
Linux OS and inside of a local Docker image). I can't (or can't find how
to) control how Binder calls , and without mounting a
volume, X11 doesn't work properly, which means PyQt5 won't work. I don't
have this issue with selenium, and it also ends up requiring fewer Linux
dependencies in binder/apt.txt. The bulk of changes in this commit are
in execute_query.py.