Adjusted .travis.yml to expand build capabilities
I found out how to get selenium to work with Chrome/Chromium as its
headless client and needed to figure out how to download Chromium
and its webdriver to a Mac for Travis tests. Had to do the same for
Firefox's driver upon realizing it also had installation problems in Travis'
Mac builds. In the end, added binder/travis_mac_installs.sh and will
address Linux in the next commit.